do not override user-specified LDFLAGS
The `LDFLAGS` assignment in `Makefile` was completely overriding the
user-specified `LDFLAGS`. I am using a non-standard location for my zlib
installation and I have set the following in my environment:

    export LDFLAGS="-L/usr/local/zlib/1.2.8-1/lib"

This does not get picked up by the old `Makefile` because `LDFLAGS` is
overridden with the `:=` assignment.

This change renames `LDFLAGS` in `Makefile` to `LDFLAGS_shared` and adds
`LDFLAGS` (the now user-specified one) to all linking stages.